% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tools.R
\name{gephi_read_edges_csv}
\alias{gephi_read_edges_csv}
\title{Read edges}
\usage{
gephi_read_edges_csv(
  path,
  source = "Source",
  target = "Target",
  directed = TRUE
)
}
\arguments{
\item{path}{The file where to read from e.g.: edges.csv or data/edges.csv}

\item{source}{What column is the source, defaults to 'Source'.}

\item{target}{What column is the target, defaults to 'Target'.}

\item{directed, }{is the graph directed or not? defaults to TRUE.}
}
\description{
This function reads in an edges csv file exported from gephi into igraph.
}
\concept{read_graphs}
