% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_tools.R
\name{gephi_write_edges_from_df}
\alias{gephi_write_edges_from_df}
\title{Write edges from a data.frame to a file}
\usage{
gephi_write_edges_from_df(dataframe, from = NULL, to = NULL, path,
  na = "", verbose = TRUE)
}
\arguments{
\item{dataframe}{which dataframe}

\item{from}{which column does the edge start}

\item{to}{which column does the edge end}

\item{path}{the file where to save to e.g.: edges.csv or data/edges.csv}

\item{na}{How to record missing values, defaults to "", nothing / empty values}

\item{verbose}{by default these functions are chatty and will tell you what they do if you do not want that, set to FALSE}
}
\value{
invisible original object so you could continue using it in a pipe if you want to
}
\description{
If you have a dataframe that represents a graph and you don't want to turn it
into an igraph/tidygraph object first, this function will turn it into a csv.
Give the column name where the edges start and the column name of the edge
end. If you don't specify it, this function expects the first column to be a
the start and second column the end point of an edge.
}
\seealso{
Other write_graphs: \code{\link{gephi_write_nodes}}
}
\concept{write_graphs}
