% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphexample.R
\docType{data}
\name{graphexample}
\alias{graphexample}
\title{a super simple graph example}
\format{
An object of class \code{igraph} of length 5.
}
\usage{
graphexample
}
\description{
I created this file to show off and test some functions.
}
\keyword{datasets}
