% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tools.R
\name{gephi_read_edges_csv}
\alias{gephi_read_edges_csv}
\title{Read edges}
\usage{
gephi_read_edges_csv(path, source = "Source", target = "Target",
  directed = TRUE)
}
\description{
Read edges
}
\concept{read_graphs}
