% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_tools.R
\name{gephi_write_edges_from_df}
\alias{gephi_write_edges_from_df}
\title{Write edges from a data.frame to a file}
\usage{
gephi_write_edges_from_df(dataframe, from = NULL, to = NULL, path,
  na = "", verbose = TRUE)
}
\arguments{
\item{dataframe}{which dataframe}

\item{from}{which column does the edge start}

\item{to}{which column does the edge end}

\item{path}{the file where to save to e.g.: edges.csv or data/edges.csv}

\item{na}{How to record missing values, defaults to "", nothing}

\item{verbose}{by default these functions are chatty and will tell you what they do if you do not want that, set to FALSE}
}
\description{
Sometimes you just want to write a data.frame to a csv.
Give the column name where the edges start and finish.
}
